+++ flexmeasures/api/__init__.py
@@ -73,7 +73,7 @@ def get_versions() -> dict:
     """
     response = {
         "message": "For these API versions a public endpoint is available, listing its service. For example: "
-        "/api/v3_0/getService. An authentication token can be requested at: "
+        "/api/v3_0. An authentication token can be requested at: "
         "/api/requestAuthToken",
         "versions": ["v3_0"],
         "flexmeasures_version": flexmeasures_version,